Update Snap SPARQL Label Utilities.txt
Changed queries so that they handle acronyms correctly. Old queries would take IRI ending in "NASAClass" and create the label "N A S A Class". With the help of ChatGPT it now has two REGEX expressions that deal with acronyms correctly. See the Wiki for more info on how ChatGPT created these.